moved chat nterfaces and made them hold array for buttons
moved chat interfaces to ChatInterfaces.tsx and made them hold MessageButtons for rasa sent buttons
