+++ PathFinder/Frontend/PathFinder-Frontend/src/Pages/Chat/Components/ChatInterfaces.tsx
@@ -1,0 +1,11 @@
+export interface MessageButton {
+  title: string;
+  payload: string;
+}
+
+export interface Message {
+  MessageNumber: number;
+  text: string;
+  role: string;
+  buttons?: MessageButton[];
+}
